+++ Sources/HColorPicker2.cpp
@@ -269,7 +269,7 @@ void HColorSlider::MouseDown(BPoint where)
 	if (knob.Contains(where))
 	{
 		BPoint oldPt(-1, -1);
-		ulong btns;
+		uint32 btns;
 		BMessage msg(msg_SliderChanged);
 		msg.AddFloat("value", fValue);
 		msg.AddPointer("sender", this);
@@ -446,7 +446,7 @@ void HColorSquare::Draw(BRect update)
 
 void HColorSquare::MouseDown(BPoint where)
 {
-	unsigned long buttons;
+	uint32 buttons;
 	int mx, my;
 	BMessage m(msg_ColorSquareChanged);
